# Improve SCAN performance by only performing expiration checks on DBs with volatile keys 