This PR optimizes scanGenericCommand by moving type filtering and expiration checks from post-processing (Step 3) to the scan callback, eliminating expensive `lookupKeyReadWithFlags()` calls and adding an optimization to skip expiration checks via dict lookup given we can now check the expiration with expiry flag in the kvobj (due to the move to the scanCallback).

Profiling data (https://pprof.me/1ac456b6d1a46b2184a5e2ef314aa0a2) showed that scanGenericCommand accounted for 2.8% of total CPU time and was a notable hotspot during SCAN-heavy workloads.

## Key optimizations:
1. **Type filtering moved to scanCallback**: Uses existing `kvobj *kv` instead of expensive lookups
2. **Expiration check optimization**: Skips `expireIfNeeded()` calls when database has no volatile keys
3. **Better cache locality**: Processes filtering during iteration rather than post-processing

Benchmarking with memtier_benchmark (1M keys, pipeline=10) showed significant performance improvements, with type-filtered scans seeing the largest gains due to eliminated lookup overhead:

| Test                  | Metric           | Baseline | Optimized (aeb229edc3f148f449077e8f7587c1e8c78b5c1d) | Δ (Improvement) |
| --------------------- | ---------------- | -------- | --------- | --------------- |
| **generic-scan**      | Ops/sec          | 800,970  | 857,274   | **+7.03%**      |
|                       | p50 Latency (ms) | 1.239    | 1.151     | **-7.11%**      |
|                       | p99 Latency (ms) | 1.695    | 1.615     | **-4.72%**      |
| **generic-scan-type** | Ops/sec          | 611,104  | 793,533   | **+29.9%**      |
|                       | p50 Latency (ms) | 1.623    | 1.247     | **-23.2%**      |
|                       | p99 Latency (ms) | 2.175    | 1.727     | **-20.6%**      |


The dramatic improvement in type-filtered scans (+30% ops/sec) demonstrates the effectiveness of eliminating expensive type lookups during scanning.

## To reproduce with benchmark suite

```bash
pip3 install redis-benchmarks-specification==0.1.267
redis-benchmarks-spec-client-runner --tests-regexp ".*-scan-" --flushall_on_every_test_start --flushall_on_every_test_end --cpuset_start_pos 2 --override-memtier-test-time 30 --benchmark_local_install --override-test-runs 3 --db_server_port <...port...> --db_server_password <...password...> --db_server_host <...host...>
```